// src/styles/types.ts
export type Breakpoints = string[]

export type MediaQueries = {
  xs: string
  sm: string
  md: string
  lg: string
  xl: string
  xxl: string
}

export type ZIndices = {
  dropdown: number
  modal: number
}

export type Shadows = {
  level: string
  active: string
  success: string
  warning: string
  focus: string
  inset: string
  tooltip: string
}

export type Radii = {
  small: string
  default: string
  card: string
  circle: string
}